+++ .gitignore
@@ -186,4 +186,4 @@ release/*
 old_app/*
 settings.env.example.bak
 TODO.md
-document/FileTree.md
+settings.env.bak
